fix(atom/slider): tests and remove useeffect bug
